fix(web): move job status dot colors to server-safe module
ProjectStatsSummary runs on the server and must not import helpers from the client job-status-badge module.
